+++ yarn-project/p2p/src/index.ts
@@ -3,6 +3,6 @@ export * from './bootstrap/bootstrap.js';
 export * from './client/index.js';
 export * from './config.js';
 export * from './mem_pools/epoch_proof_quote_pool/index.js';
-export * from './service/index.js';
+export * from './services/index.js';
 export * from './mem_pools/tx_pool/index.js';
 export * from './tx_validator/index.js';
