--- yarn-project/p2p/src/service/index.ts
@@ -1,2 +1,0 @@
-export * from './service.js';
-export * from './libp2p_service.js';
