+++ yarn-project/p2p/src/services/encoding.ts
