+++ yarn-project/p2p/src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './service.js';
+export * from './libp2p/libp2p_service.js';
