+++ yarn-project/p2p/src/services/reqresp/rate_limiter/rate_limiter.test.ts
@@ -2,8 +2,8 @@ import { jest } from '@jest/globals';
 import { type PeerId } from '@libp2p/interface';
 import { type MockProxy, mock } from 'jest-mock-extended';
 
+import { PeerErrorSeverity } from '../../peer-scoring/peer_scoring.js';
 import { type PeerManager } from '../../peer_manager.js';
-import { PeerErrorSeverity } from '../../peer_scoring.js';
 import { PING_PROTOCOL, type ReqRespSubProtocolRateLimits, TX_REQ_PROTOCOL } from '../interface.js';
 import { RequestResponseRateLimiter } from './rate_limiter.js';
